feat(docs): Add header to pb.cls, fix python - Add header command to pb.cls (title sets header and is common to both   ug and pb) - Fix circular dependency on python mkregs and verilog2tex scripts
