+++ document/pb/pb.cls
@@ -53,6 +53,7 @@
 %}
 
 \newcommand{\category}[1]{\def\@category{#1}}
+\newcommand{\header}[1]{\def\@header{#1}}
 \renewcommand{\familydefault}{\sfdefault}
 
 \renewcommand{\maketitle}{
